+++ spec/friendly_shipping/carrier_spec.rb
@@ -6,6 +6,7 @@ RSpec.describe FriendlyShipping::Carrier do
   it { is_expected.to respond_to :id }
   it { is_expected.to respond_to :name }
   it { is_expected.to respond_to :code }
+  it { is_expected.to respond_to :scac }
   it { is_expected.to respond_to :shipping_methods }
   it { is_expected.to respond_to :balance }
   it { is_expected.to respond_to :data }